+++ app/lib/methods/subscriptions/rooms.js
@@ -1,6 +1,8 @@
+import Random from 'react-native-meteor/lib/Random';
 import database from '../../realm';
 import { merge } from '../helpers/mergeSubscriptionsRooms';
 import protectedFunction from '../helpers/protectedFunction';
+import messagesStatus from '../../../constants/messagesStatus';
 
 export default async function subscribeRooms(id) {
 	const subscriptions = Promise.all([
@@ -25,7 +27,9 @@ export default async function subscribeRooms(id) {
 		}, 5000);
 	};
 
-	if (this.ddp) {
+	if (!this.ddp && this._login) {
+		loop();
+	} else {
 		this.ddp.on('logged', () => {
 			clearTimeout(timer);
 			timer = false;
@@ -47,8 +51,8 @@ export default async function subscribeRooms(id) {
 			const [, ev] = ddpMessage.fields.eventName.split('/');
 			if (/subscriptions/.test(ev)) {
 				const tpm = merge(data);
-				return database.write(() => {
-					merge(tpm);
+				database.write(() => {
+					database.create('subscriptions', tpm, true);
 				});
 			}
 			if (/rooms/.test(ev) && type === 'updated') {
@@ -56,6 +60,25 @@ export default async function subscribeRooms(id) {
 				database.write(() => {
 					merge(sub, data);
 				});
+			}
+			if (/message/.test(ev)) {
+				const [args] = ddpMessage.fields.args;
+				const _id = Random.id();
+				const message = {
+					_id,
+					rid: args.rid,
+					msg: args.msg,
+					ts: new Date(),
+					_updatedAt: new Date(),
+					status: messagesStatus.SENT,
+					u: {
+						_id,
+						username: 'rocket.cat'
+					}
+				};
+				requestAnimationFrame(() => database.write(() => {
+					database.create('messages', message, true);
+				}));
 			}
 		}));
 	}
