+++ e2e/06-createroom.spec.js
@@ -6,6 +6,7 @@ const data = require('./data');
 
 describe('Create room screen', () => {
 	before(async() => {
+		await device.reloadReactNative(); // TODO: remove this after fix logout subscription
 		await element(by.id('rooms-list-view-create-channel')).tap();
 		await waitFor(element(by.id('select-users-view'))).toBeVisible().withTimeout(2000);
 	});
